Split TR files based on main algorithm
Split TR decomposition and testing files based on type of algorithm (SVD or ALS).
